WIP: Continue adding collation information to gporca
Determination was made that we need to add the collation information
to the CColRef objects in order to pass collation information through
to operatiors, in addition to the changes I made before. This follows
with the changes made to add typmod fields to the codebase in
ecba302c3e83e62e998f1fdaed5bb3c09f010a6

Next steps: Continue adding collation field to related objects.