WIP: Use an InvalidOID to create column references
I'm skeptical that this will work, but this commit will contain the
changes that are related to the InvalidOID changes.

TODO: I have added NULL as a placeholder for several values here in an
attempt to find more compiler issues. I need to return to this commit
as soon as I pair with QP so they can help me figure out where to put
the InvalidOID default variable.